# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Framework MCP API - Pure Data Provider",
    "description": "Microsoft Copilot-compatible API providing authoritative CIS Controls Framework v8.1 data (153 safeguards). Pure Data Provider architecture empowers LLMs with structured safeguards data for sophisticated, context-aware vendor capability analysis.",
    "version": "2.6.1",
    "contact": {
      "name": "Framework MCP Support",
      "url": "https://github.com/therealcybermattlee/FrameworkMCP"
    },
    "license": {
      "name": "CC-BY-4.0",
      "url": "https://creativecommons.org/licenses/by/4.0/"
    }
  },
  "servers": [
    {
      "url": "https://your-deployment.example.com",
      "description": "Production deployment"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/safeguards/{safeguardId}": {
      "get": {
        "summary": "Get Safeguard Details",
        "description": "Get detailed information about a specific CIS safeguard including governance elements, core requirements, sub-taxonomical elements, and implementation suggestions",
        "operationId": "getSafeguardDetails",
        "tags": ["Safeguards"],
        "parameters": [
          {
            "name": "safeguardId",
            "in": "path",
            "required": true,
            "description": "CIS safeguard ID (e.g., '1.1', '5.1', '12.8')",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+\\.[0-9]+$",
              "example": "1.1"
            }
          },
          {
            "name": "include_examples",
            "in": "query",
            "required": false,
            "description": "Include implementation examples in response",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Safeguard details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeguardDetails"
                }
              }
            }
          },
          "400": {
            "description": "Invalid safeguard ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Safeguard not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SafeguardDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1.1"
          },
          "title": {
            "type": "string",
            "example": "Establish and Maintain Detailed Enterprise Asset Inventory"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the safeguard requirements"
          },
          "implementationGroup": {
            "type": "string",
            "enum": ["IG1", "IG2", "IG3"],
            "example": "IG1"
          },
          "governanceElements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Governance requirements that MUST be met (Orange elements)"
          },
          "coreRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Core 'what' requirements of the safeguard (Green elements)"
          },
          "subTaxonomicalElements": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sub-taxonomical components (Yellow elements)"
          },
          "implementationSuggestions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Implementation methods and suggestions (Gray elements)"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "string",
            "description": "Additional error details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}